requirements.txt
python-dotenv>=1.0.0
web3>=6.20
py-clob-client>=0.5.0
requests>=2.31
tenacity>=8.2.0
orjson>=3.10
pydantic>=2.5
